Refactor/service controller router update

+++ src/services/category_service.js
@@ -2,25 +2,41 @@ import db from "../models";
 
 export const addCategory = async (data) => {
 	try {
-		const response = await db.Category.create(data);
-		if (response) {
-			return {
-				message: "Create category successfully",
-				data: {
-					id: response.id,
-					name: response.name,
-					img: response.img,
-					description: response.description,
-				},
-			};
+		const existingCate = await db.Category.findOne({
+			where: {
+				name: data.name,
+			},
+		});
+		if (existingCate) {
+			throw { status: 400, message: "Danh mục đã tồn tại!" };
 		}
+		const response = await db.Category.create(data);
+		return { data: response };
 	} catch (error) {
-		console.error(error);
-		throw new Error("Database Error");
+		console.log(error);
+		throw error;
 	}
 };
 export const updateCategory = async (data) => {
 	try {
+		const foundCategory = await db.Category.findOne({
+			where: {
+				id: param,
+			},
+		});
+		if (!foundCategory) {
+			throw {
+				status: 404,
+				message: "Danh mục không tồn tại, cập nhật không thành công!",
+			};
+		}
+
+		if (foundCategory.name === data.name) {
+			throw {
+				status: 400,
+				message: "Tên danh mục đã tồn tại, cập nhật không thành công!",
+			};
+		}
 		const response = await db.Category.update(
 			{ name: data.name, img: data.img, description: data.description },
 			{
@@ -34,30 +50,39 @@ export const updateCategory = async (data) => {
 		});
 		if (response) {
 			return {
-				message: "update category successfully",
+				message: "Cập nhật thành công !",
 				data: findData,
 			};
 		}
 	} catch (error) {
 		console.error(error);
-		throw new Error("Database Error");
+		throw error;
 	}
 };
 export const deleteCategory = async (id) => {
 	try {
-		const response = await db.Category.destroy({
+		const foundCategory = await db.Category.findOne({
+			where: {
+				id: param,
+			},
+		});
+		if (!foundCategory) {
+			throw {
+				status: 404,
+				message: "Danh mục không tồn tại, xóa không thành công!",
+			};
+		}
+		await db.Category.destroy({
 			where: {
 				id: id,
 			},
 		});
-		if (response) {
-			return {
-				message: "Delete successfully",
-			};
-		}
+		return {
+			message: "Xóa thành công",
+		};
 	} catch (error) {
 		console.error(error);
-		throw new Error("Database Error");
+		throw error;
 	}
 };
 export const getCategory = async () => {
@@ -65,14 +90,12 @@ export const getCategory = async () => {
 		const response = await db.Category.findAll({
 			attributes: ["id", "name", "img", "description"],
 		});
-		if (response) {
-			return {
-				data: response,
-			};
-		}
+		return {
+			data: response,
+		};
 	} catch (error) {
 		console.error(error);
-		throw new Error("Database Error");
+		throw error;
 	}
 };
 export const getOneCategory = async (id) => {
@@ -80,13 +103,11 @@ export const getOneCategory = async (id) => {
 		const response = await db.Category.findByPk(id, {
 			attributes: ["id", "name", "img", "description"],
 		});
-		if (response) {
-			return {
-				data: response,
-			};
-		}
+		return {
+			data: response,
+		};
 	} catch (error) {
 		console.error(error);
-		throw new Error("Database Error");
+		throw error;
 	}
 };
